+++ feersum_nlu/api/__init__.py
@@ -9,7 +9,6 @@ from feersum_nlu.api.dashboard_api import DashboardApi
 from feersum_nlu.api.data_objects_api import DataObjectsApi
 from feersum_nlu.api.date_parsers_api import DateParsersApi
 from feersum_nlu.api.duckling_entity_extractors_api import DucklingEntityExtractorsApi
-from feersum_nlu.api.emotion_classifiers_api import EmotionClassifiersApi
 from feersum_nlu.api.faq_matchers_api import FaqMatchersApi
 from feersum_nlu.api.health_api import HealthApi
 from feersum_nlu.api.image_classifiers_api import ImageClassifiersApi
@@ -22,8 +21,6 @@ from feersum_nlu.api.params_api import ParamsApi
 from feersum_nlu.api.regex_entity_extractors_api import RegexEntityExtractorsApi
 from feersum_nlu.api.sentence_translators_api import SentenceTranslatorsApi
 from feersum_nlu.api.sentiment_detectors_api import SentimentDetectorsApi
-from feersum_nlu.api.sim_word_entity_extractors_api import SimWordEntityExtractorsApi
 from feersum_nlu.api.synonym_entity_extractors_api import SynonymEntityExtractorsApi
 from feersum_nlu.api.text_classifiers_api import TextClassifiersApi
 from feersum_nlu.api.text_datasets_api import TextDatasetsApi
-from feersum_nlu.api.text_pipelines_api import TextPipelinesApi