+++ test/test_duckling_extractor.py
@@ -69,7 +69,8 @@ class TestDucklingExtractor(unittest.TestCase):
             entity_list = api_response
             if len(entity_list) > 0:
                 entity = entity_list[0]
-                self.assertTrue((entity.value == '2017-01-05 15:00:00'))
+                print(f"entity.value = {entity.value}")
+                self.assertTrue((entity.value == '2017-01-05T15:00:00.000'))
             else:
                 self.assertTrue(False)