+++ test/test_duckling_extractor_perf.py
@@ -71,7 +71,7 @@ class TestDucklingExtractorPerf(unittest.TestCase):
             entity_list = api_response
             if len(entity_list) > 0:
                 entity = entity_list[0]
-                self.assertTrue((entity.value == '2017-01-05 15:00:00'))
+                self.assertTrue((entity.value == '2017-01-05T15:00:00.000'))
             else:
                 self.assertTrue(False)
 
@@ -87,7 +87,7 @@ class TestDucklingExtractorPerf(unittest.TestCase):
                 entity_list = api_response
                 if len(entity_list) > 0:
                     entity = entity_list[0]
-                    self.assertTrue((entity.value == '2017-01-05 15:00:00'))
+                    self.assertTrue((entity.value == '2017-01-05T15:00:00.000'))
                 else:
                     self.assertTrue(False)