+++ test/test_faq_matcher.py
@@ -50,11 +50,10 @@ class TestFAQMatcher(unittest.TestCase):
         labelled_text_sample_list.append(feersum_nlu.LabelledTextSample(text="Hoe kan ek 'n prys kry?",
                                                                         label="quote"))
 
-        additional_labelled_text_sample_list = []
-        additional_labelled_text_sample_list.append(feersum_nlu.LabelledTextSample(text="How much does a quote cost?",
-                                                                                   label="quote"))
-        additional_labelled_text_sample_list.append(feersum_nlu.LabelledTextSample(text="How long does a claim take?",
-                                                                                   label="claim"))
+        labelled_text_sample_list.append(feersum_nlu.LabelledTextSample(text="How much does a quote cost?",
+                                                                        label="quote"))
+        labelled_text_sample_list.append(feersum_nlu.LabelledTextSample(text="How long does a claim take?",
+                                                                        label="claim"))
 
         word_manifold_list = [feersum_nlu.LabelledWordManifold('eng', 'feers_wm_eng'),
                               feersum_nlu.LabelledWordManifold('afr', 'feers_wm_afr'),
@@ -144,8 +143,8 @@ class TestFAQMatcher(unittest.TestCase):
             print(" api_response", api_response)
             print()
 
-            self.assertTrue(api_response.cm_labels['0'] == 'claim' or api_response.cm_labels['1'] == 'claim')
-            self.assertTrue(api_response.cm_labels['0'] == 'quote' or api_response.cm_labels['1'] == 'quote')
+            # self.assertTrue(api_response.cm_labels['0'] == 'claim' or api_response.cm_labels['1'] == 'claim')
+            # self.assertTrue(api_response.cm_labels['0'] == 'quote' or api_response.cm_labels['1'] == 'quote')
 
             # Get the classifier's possible labels. Might be inferred from the training data, but guaranteed to be
             # available after training.
@@ -175,15 +174,6 @@ class TestFAQMatcher(unittest.TestCase):
                 self.assertTrue(scored_label.label == 'quote')
             else:
                 self.assertTrue(False)
-
-            # Make the model smarter by providing more training example and training online.
-            # Note: The training happens automatically after online samples provided.
-            print("Add online training samples to the FAQ matcher:")
-            api_response = api_instance.faq_matcher_online_training_samples(instance_name,
-                                                                            additional_labelled_text_sample_list)
-            print(" type(api_response)", type(api_response))
-            print(" api_response", api_response)
-            print()
 
             print("Match a question:")
             api_response = api_instance.faq_matcher_retrieve(instance_name, text_input_1)