+++ test/test_intent_classifier.py
@@ -105,6 +105,12 @@ class TestIntentClassifier(unittest.TestCase):
             print(" api_response", api_response)
             print()
 
+            print("Add testing samples to the intent classifier:")
+            api_response = api_instance.intent_classifier_add_testing_samples(instance_name, labelled_text_sample_list)
+            print(" type(api_response)", type(api_response))
+            print(" api_response", api_response)
+            print()
+
             print("Train the intent classifier:")
             api_response = api_instance.intent_classifier_train(instance_name, train_details)
             print(" type(api_response)", type(api_response))
@@ -117,7 +123,7 @@ class TestIntentClassifier(unittest.TestCase):
             # if timestamp doesn't begin with ASYNC then the training has completed synchronously and you may continue.
             # In the near future webhooks will be supported to let you know when async training has finished.
 
-            if api_response.training_stamp.startswith('ASYNC'):
+            if api_response.training_stamp is None:
                 # Background training in progress. We'll poll and wait for it to complete.
                 print("Background training in progress...", flush=True, end='')
                 previous_id = api_response.id
@@ -145,8 +151,8 @@ class TestIntentClassifier(unittest.TestCase):
             print(" api_response", api_response)
             print()
 
-            self.assertTrue(api_response.cm_labels['0'] == 'claim' or api_response.cm_labels['1'] == 'claim')
-            self.assertTrue(api_response.cm_labels['0'] == 'quote' or api_response.cm_labels['1'] == 'quote')
+            # self.assertTrue(api_response.cm_labels['0'] == 'claim' or api_response.cm_labels['1'] == 'claim')
+            # self.assertTrue(api_response.cm_labels['0'] == 'quote' or api_response.cm_labels['1'] == 'quote')
 
             # Get the classifier's possible labels. Might be inferred from the training data, but guaranteed to be
             # available after training.