+++ test/test_text_classifier_usent.py
@@ -105,12 +105,11 @@ class TestTextClassifier(unittest.TestCase):
             print()
 
             # TRAINING:
-            # If timestamp begins with 'ASYNC...' the the training is running in the background and you need to poll until
+            # If timestamp is missing then the training is running in the background and you need to poll until
             # the model ID has updated.
-            # if timestamp doesn't begin with ASYNC then the training has completed synchronously and you may continue.
             # In the near future webhooks will be supported to let you know when async training has finished.
 
-            if api_response.training_stamp.startswith('ASYNC'):
+            if api_response.training_stamp is None:
                 # Background training in progress. We'll poll and wait for it to complete.
                 print("Background training in progress...", flush=True, end='')
                 previous_id = api_response.id